delivery service realization

+++ src/main/java/com/upp/apteka/service/impl/DeliveryServiceImpl.java
@@ -16,8 +16,8 @@ import com.upp.apteka.bo.Medicine;
 import com.upp.apteka.bo.Pharmacy;
 import com.upp.apteka.bo.PharmacyMedicine;
 import com.upp.apteka.repository.DeliveryRepository;
-import com.upp.apteka.repository.PharmacyRepository;
 import com.upp.apteka.service.DeliveryService;
+import com.upp.apteka.service.PharmacyService;
 
 @Service("deliveryService")
 @Transactional
@@ -27,15 +27,20 @@ public class DeliveryServiceImpl implements DeliveryService {
 	DeliveryRepository deliveryRepository;
 
 	@Autowired
-	PharmacyRepository pharmacyRepository;
+	PharmacyService pharmacyService;
 
 	public List<Delivery> getAllDeliveries(int offset) {
-		
+
 		return deliveryRepository.getAll(offset, Constants.LIMIT);
 	}
 
-	public Delivery readDelivery(Long id) {
-		
+	public List<Delivery> getAllPharmacyDeliveries(Long pharmacyId, int offset) {
+
+		return deliveryRepository.findPharmacyDeliveries(pharmacyId, offset, Constants.LIMIT);
+	}
+
+	public Delivery getDelivery(Long id) {
+
 		return deliveryRepository.read(id);
 	}
 
@@ -48,7 +53,7 @@ public class DeliveryServiceImpl implements DeliveryService {
 		List<DeliveryMedicine> delMeds = delivery.getDeliveryMedicines();
 
 		// get delivery pharmacy
-		Pharmacy pharmacy = pharmacyRepository.read(delivery.getPharmacy().getId());
+		Pharmacy pharmacy = pharmacyService.getPharmacy(delivery.getPharmacy().getId());
 
 		// get all PharmacyMedicines by pharmacyId
 		List<PharmacyMedicine> pharmMeds = pharmacy.getPharmacyMedicines();
@@ -76,65 +81,64 @@ public class DeliveryServiceImpl implements DeliveryService {
 			}
 		}
 
-		pharmacyRepository.update(pharmacy);
+		pharmacyService.updatePharmacy(pharmacy);
 
 	}
-	
+
 	public void updateDelivery(Delivery delivery) {
-		
+
 		deliveryRepository.update(delivery);
 	}
 
 	public boolean deleteDelivery(Long id) {
-		/*
-		 * Delivery delivey = deliveryRepository.read(id); Pharmacy pharmacy =
-		 * pharmacyRepository.read(delivey.getPharmacy().getId());
-		 * 
-		 * List<PharmacyMedicine> pharmMeds = pharmacy.getPharmacyMedicines();
-		 * 
-		 * // get delivery medicines List<DeliveryMedicine> delMeds =
-		 * delivey.getDeliveryMedicines();
-		 * 
-		 * for (DeliveryMedicine delMed : delMeds) { if
-		 * (pharmacyMedicineCheck(pharmMeds, delMed)) {
-		 * 
-		 * for (PharmacyMedicine pharmMed : pharmMeds) { if
-		 * (pharmMed.getMedicine().getId() == delMed.getMedicine().getId()) {
-		 * 
-		 * if(pharmMed.getPackQuantity() -
-		 * delMed.getBoxQuantity()*delMed.getMedicine().getQuantityPerBox()
-		 * >=0){ pharmMed.setPackQuantity(pharmMed.getPackQuantity() -
-		 * delMed.getBoxQuantity()*delMed.getMedicine().getQuantityPerBox()); }
-		 * } }
-		 * 
-		 * } else { // create new pharmacy medicine pharmacyMedicine = new
-		 * PharmacyMedicine();
-		 * pharmacyMedicine.setMedicine(delMed.getMedicine());
-		 * pharmacyMedicine.setPharmacy(delivery.getPharmacy());
-		 * pharmacyMedicine.setPackQuantity(generatePMPackQuantity(delMed));
-		 * pharmacyMedicine.setPackPrice(generatePMPackPrice(delivery.
-		 * getPharmacy(), delMed.getMedicine()));
-		 * 
-		 * pharmMeds.add(pharmacyMedicine); } }
-		 * 
-		 */
+		
+		if(deliveryRepository.checkIfDeliveryMedicineSold(id)){
+			return false;
+		}else{
+			//update pharmacy medicines quntity
+			Delivery delivery = deliveryRepository.read(id);
+			
+			// get delivery pharmacy
+			Pharmacy pharmacy = pharmacyService.getPharmacy(delivery.getPharmacy().getId());
 
-		return deliveryRepository.delete(id);
+			// get all PharmacyMedicines by pharmacyId
+			List<PharmacyMedicine> pharmMeds = pharmacy.getPharmacyMedicines();
 
+			List<DeliveryMedicine> delMeds = delivery.getDeliveryMedicines();
+			
+			for (DeliveryMedicine delMed : delMeds) {
+				if (pharmacyMedicineCheck(pharmMeds, delMed)) {
+
+					for (PharmacyMedicine pharmMed : pharmMeds) {
+						if (pharmMed.getMedicine().getId().equals(delMed.getMedicine().getId())) {
+							pharmMed.setPackQuantity(pharmMed.getPackQuantity() - generatePMPackQuantity(delMed));
+						}
+					}
+
+				} 
+				}	
+			pharmacyService.updatePharmacy(pharmacy);
+			return deliveryRepository.delete(id);
+		}
 	}
 
 	// Delivery repo specific requests
 
-		public List<Delivery> findPharmacyDeliveriesByPeriod(Date from, Date to, Long pharmacyId, int offset) {
-			return deliveryRepository.findPharmacyDeliveriesByPeriod(from, to, pharmacyId, offset, Constants.LIMIT);
-		}
+	public List<Delivery> getPharmacyDeliveriesByPeriod(Date from, Date to, Long pharmacyId, int offset) {
+		return deliveryRepository.findPharmacyDeliveriesByPeriod(from, to, pharmacyId, offset, Constants.LIMIT);
+	}
 
-		public List<Delivery> findPharmacyMedicineDeliveriesByPeriod(Date from, Date to, Long pharmacyId, Long medicineId,
-				int offset) {
-			return deliveryRepository.findPharmacyMedicineDeliveriesByPeriod(from, to, pharmacyId, medicineId, offset,
-					Constants.LIMIT);
-		}
+	public List<Delivery> getPharmacyMedicineDeliveriesByPeriod(Date from, Date to, Long pharmacyId, Long medicineId,
+			int offset) {
+		return deliveryRepository.findPharmacyMedicineDeliveriesByPeriod(from, to, pharmacyId, medicineId, offset,
+				Constants.LIMIT);
+	}
+	
+	public boolean checkIfDeliveryMedicineSold(Long deliveryId){
 		
+		return false;
+	}
+
 	private int generatePMPackQuantity(DeliveryMedicine delMed) {
 
 		return delMed.getBoxQuantity() * delMed.getMedicine().getQuantityPerBox();
@@ -158,13 +162,12 @@ public class DeliveryServiceImpl implements DeliveryService {
 				MathContext.DECIMAL64);
 
 		return pricePerPackWithVAT;
-	}	
+	}
 
 	private boolean pharmacyMedicineCheck(List<PharmacyMedicine> pharmMeds, DeliveryMedicine delMed) {
 
 		for (PharmacyMedicine pharmMed : pharmMeds) {
 			// if medicine exists in the pharmacy
-
 			System.out.println("delMed id: " + delMed.getMedicine().getId());
 			System.out.println("pharmMed id: " + pharmMed.getMedicine().getId());
 			if (delMed.getMedicine().getId().equals(pharmMed.getMedicine().getId())) {
@@ -173,7 +176,5 @@ public class DeliveryServiceImpl implements DeliveryService {
 		}
 		return false;
 	}
-
-	
 
 }
